multimodal-agent/requirements.txt
chromadb==1.5.5
langchain==1.2.10
langchain-chroma==1.1.0
langchain-community==0.4.1
langchain-core==1.2.13
langchain-groq==1.1.2
langchain-huggingface==1.2.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.0
langgraph==1.0.8
openai==2.21.0
pillow==11.3.0
pyside6==6.9.3
pypdf==6.9.1
python-dotenv==1.2.1
sentence-transformers==5.3.0
torch==2.10.0
